+++ src/page/components/chat-step/question-result/question-result.tsx
@@ -1,12 +1,10 @@
-import PresetButton from '@/components/preset-button/preset-button';
-import PRESET_BUTTON from '@/constants/preset-buttons';
 import { AdmissionPresetType } from '@/constants/preset-buttons';
 import useAdmissionQuestionResult from '@/hooks/use-admission-question-result';
+import PresetButtons from '@/page/components/preset-buttons/preset-buttons';
 import useAdmissionStore from '@/stores/store/admission-store';
 import useMessagesStore from '@/stores/store/message-store';
 import { AdmissionType } from '@/types/admission-type';
-import { ChatSteps } from '@/types/chat';
-import { apiEventGATrigger } from '@/utils/ga-trigger';
+import { ChatSteps } from '@/types/steps';
 import getQuestionPrompt from '@/utils/get-question-prompt';
 
 interface QuestionResultProps {
@@ -28,7 +26,7 @@ function QuestionResult({ admissionType, admissionCategory, question, changeStep
     questionType: 'detail',
   });
 
-  const selectQuestion = (question: AdmissionPresetType) => {
+  const handleQuestionSelect = (question: AdmissionPresetType) => {
     refetch();
     setMessages([
       {
@@ -41,45 +39,7 @@ function QuestionResult({ admissionType, admissionCategory, question, changeStep
 
   if (isLoading) return <div>Loading</div>;
 
-  return (
-    <div className="mt-2 flex w-full justify-end">
-      <div className="flex w-72 flex-wrap justify-end gap-2">
-        <PresetButton
-          onClick={() => {
-            changeStep('질문 출처 결과');
-            setMessages([{ role: 'user', message: '🙋‍♂️ 어디에서 볼 수 있나요?' }]);
-            apiEventGATrigger({ category: 'preset button click', action: 'click', label: '출처확인하기', value: 1 });
-          }}
-        >
-          🙋‍♂️ 어디에서 볼 수 있나요?
-        </PresetButton>
-        {PRESET_BUTTON.map((question) => (
-          <PresetButton
-            onClick={() => {
-              selectQuestion(question);
-              apiEventGATrigger({
-                category: 'preset button click',
-                action: 'click',
-                label: `${question}질문`,
-                value: 1,
-              });
-            }}
-            key={question.label}
-          >
-            {question.label}
-          </PresetButton>
-        ))}
-        <PresetButton
-          onClick={() => {
-            window.location.reload();
-            apiEventGATrigger({ category: 'preset button click', action: 'click', label: '조건재설정', value: 1 });
-          }}
-        >
-          조건 재설정
-        </PresetButton>
-      </div>
-    </div>
-  );
+  return <PresetButtons changeStep={changeStep} handleQuestionSelect={handleQuestionSelect} />;
 }
 
 export default QuestionResult;
